tests/wpt/tests/css/css-align/parsing/grid-row-gap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing grid-row-gap with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align/#grid-gap-legacy">
<link rel="author" title="Takuya Kurimoto" href="mailto:takuya004869@gmail.com">
<meta name="assert" content="grid-row-gap supports only the grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("grid-row-gap", "auto");
test_invalid_value("grid-row-gap", "-10px");

test_invalid_value("grid-row-gap", "10px 20%");
test_invalid_value("grid-row-gap", "normal 10px");
</script>
</body>
</html>
